Updating deprecated modules and functions.
Removing references to the 'patterns' module, that was deprecated in Django 1.8.
Changing the references from the module 'django.core.urlresolvers' to 'django.urls' because the first one was removed in Django 2.0.
Changing the references from the module 'django.conf.urls' to 'django.urls' because the first one it's soon to be deprecated.
Changing the references from the 'url' function to 're_path' because the first one it's soon to be deprecated.